Arreglados varios warnings

+++ MasterUAB/Graphics/Effects/EffectManager.cpp
@@ -109,7 +109,7 @@ void CEffectManager::SetLightConstants(unsigned int IdLight, CLight *Light)
 
 void CEffectManager::SetLightsConstants(unsigned int MaxLights)
 {
-	for (int i = 0; i < MaxLights; i++)
+	for (size_t i = 0; i < MaxLights; i++)
 	{
 		SetLightConstants(i, UABEngine.GetLightManager()->GetResourceById(i));
 	}
